+++ packages/sdk/src/chains/solana/actions/index.ts
@@ -5,6 +5,5 @@
  */
 
 export * from './redeem';
-export * from './redeemForBtc';
 export * from './stake';
 export * from './unstake';